requirements.txt
# Core LLM and AI libraries
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.6.1

# Vector databases and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Web framework and API
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1

# Data processing
pandas==2.1.4
numpy==1.24.3
pydantic==2.5.0

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
sqlite3

# Utilities
tqdm==4.66.1
rich==13.7.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
